# swagger.yaml
openapi: 3.0.0
info:
  title: Fitness API
  description: |
    API RESTful complète pour l'application mobile Fitness.
    Supporte gestion des utilisateurs, courses, séances, pas, poids, amis, messages et badges.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fitnesssapi.com
  license:
    name: MIT
servers:
  - url: https://fitnesssapi.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentification par Bearer Token (JWT)
  schemas:
...
    RunningSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          example: 3600
        distanceKm:
          type: number
          example: 10.5
        avgSpeedKmh:
          type: number
          example: 10.5
        maxSpeedKmh:
          type: number
          example: 12
        caloriesBurned:
          type: number
          example: 750
        elevationGainM:
          type: number
          example: 150
        avgHeartRate:
          type: integer
          example: 140
        maxHeartRate:
          type: integer
          example: 165
        weather:
          type: string
          nullable: true
          example: sunny
        temperatureC:
          type: number
          nullable: true
          example: 22
        notes:
          type: string
          nullable: true
        isCompleted:
          type: boolean
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        gpsPoints:
          type: array
          items:
            $ref: '#/components/schemas/GPSPoint'
    RunningSessionCreate:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationSeconds:
          type: integer
        distanceKm:
          type: number
        avgSpeedKmh:
          type: number
        maxSpeedKmh:
          type: number
        caloriesBurned:
          type: number
        elevationGainM:
          type: number
        elevationLossM:
          type: number
        avgHeartRate:
          type: integer
        maxHeartRate:
          type: integer
        weather:
          type: string
        temperatureC:
          type: number
        notes:
          type: string
        isCompleted:
          type: boolean
        gpsPoints:
          type: array
          items:
            $ref: '#/components/schemas/GPSPointInput'
    GPSPoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        latitude:
          type: number
          example: 48.8566
        longitude:
          type: number
          example: 2.3522
        altitudeM:
          type: number
          nullable: true
        speedKmh:
          type: number
          nullable: true
        accuracyM:
          type: number
          nullable: true
        recordedAt:
          type: string
          format: date-time
    GPSPointInput:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        altitudeM:
          type: number
        speedKmh:
          type: number
        accuracyM:
          type: number
        recordedAt:
          type: string
          format: date-time
    RunningStats:
      type: object
      properties:
        totalSessions:
          type: integer
          example: 42
        totalDistanceKm:
          type: number
          example: 525.5
        totalDurationSeconds:
          type: integer
          example: 151200
        totalCaloriesBurned:
          type: number
          example: 31500

    # ===== Workout Schemas =====
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Code d'erreur
          example: VALIDATION_ERROR
        message:
          type: string
          description: Message d'erreur lisible
          example: Email or password incorrect
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ===== Sync Schemas =====
...
paths:
...
  /running/stats/summary:
    get:
      tags:
        - Running
      summary: Statistiques généraldes courses
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningStats'
  /running:
    get:
      tags:
        - Running
      summary: Lister les sessions de course
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Liste des sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Running
      summary: Créer une session de course
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunningSessionCreate'
      responses:
        '201':
          description: Session créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningSession'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /running/{id}:
    get:
      tags:
        - Running
      summary: Récupérer une session
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de la session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningSession'
        '404':
          description: Session non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Running
      summary: Supprimer une session
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ===== WORKOUTS =====
...